+++ core/test/gas/LiquidatorGas.t.sol
@@ -211,7 +211,7 @@ contract LiquidatorGasTest is Test, IManager, ILiquidator {
     }
 
     // IManager
-    function callback(bytes calldata data, address, uint144) external returns (uint144 positions) {
+    function callback(bytes calldata data, address, uint208) external returns (uint208 positions) {
         require(msg.sender == address(account));
 
         (Action action, uint256 amount0, uint256 amount1) = abi.decode(data, (Action, uint256, uint256));
